# Three-tier fast-fail URL validation with playlist rejection