--- a/.claude/commands/coderabbit2.md
+++ b/.claude/commands/coderabbit2.md
@@ -1,0 +1,26 @@
+---
+description: Analyse CodeRabbit AI comment and recommend if to proceed
+argument-hint: "[code-rabbit-comment]"
+---
+
+# Recommend if to action CodeRabbit AI comment
+
+This branch as an open PR which CodeRabbit AI has provided some comments. CodeRabbit comments are NOT always sane or relevant as CodeRabbit does not have the full context of this code base. **Your task** is to recommend if we should action the comment.
+
+## Comment to Review
+
+```
+$ARGUMENTS
+```
+
+## Your Task Process
+
+1. Explain comment simply
+2. Analyse the surrounding relevant code to determine:
+   - Is comment is sane and valuable to action in context of our codebase?
+   - does this improve code clarity, elegance, simplicity, good practices?
+   - consider complexity of change versus value, is it pragmatic and worth it?
+   - lightly consider CLI application may oneday be a API Service, does it detract?
+3. Make your recommendation and provide reasoning
+
+Include emoji's in your response for readability.