… README.md
… docs/refactor-todo/exceptions/test_file.md
… docs/refactor-todo/exceptions/test_url.md
… docs/refactor-todo/ideas.md